services/extractor/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiokafka>=0.9.0
spacy>=3.7.0
sentence-transformers>=2.2.0
python-dateutil>=2.8.0
numpy>=1.24.0
